requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.18
jinja2==3.1.5
openai==1.58.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.60
langchain-core==0.3.28
faiss-cpu==1.9.0.post1
duckduckgo-search==7.2.1
tiktoken==0.8.0
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
